integrating with fakestore api

+++ src/main/java/com/yash/scaler/productservice8aug/model/Category.java
@@ -6,6 +6,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class Category {
-    private Long id;
+    private Integer id;
     private String title;
+
 }
